+++ ilert/data_source_escalation_policy.go
@@ -38,9 +38,9 @@ func dataSourceEscalationPolicyRead(ctx context.Context, d *schema.ResourceData,
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for escalation policy with name '%s' to be read", searchName))
+				return resource.RetryableError(fmt.Errorf("waiting for escalation policy with name '%s' to be read, error: %s", searchName, err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not read an escalation policy with name: %s", searchName))
+			return resource.NonRetryableError(fmt.Errorf("could not read an escalation policy with name: %s, error: %s", searchName, err.Error()))
 		}
 
 		found := resp.EscalationPolicy