+++ ilert/data_source_user_phone_number_contact.go
@@ -68,9 +68,9 @@ func dataSourceUserPhoneNumberContactRead(ctx context.Context, d *schema.Resourc
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for user contact with phone number '%s' to be read", searchTarget))
+				return resource.RetryableError(fmt.Errorf("waiting for user contact with phone number '%s' to be read, error: %s", searchTarget, err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not read a user contact with phone number: %s", searchTarget))
+			return resource.NonRetryableError(fmt.Errorf("could not read a user contact with phone number: %s, error: %s", searchTarget, err.Error()))
 		}
 
 		found := resp.UserPhoneNumberContact