+++ ilert/resource_alert_source.go
@@ -915,7 +915,7 @@ func resourceAlertSourceCreate(ctx context.Context, d *schema.ResourceData, m in
 		return diag.FromErr(err)
 	}
 	if result == nil || result.AlertSource == nil {
-		log.Printf("[ERROR] Creating ilert alert source error: empty response ")
+		log.Printf("[ERROR] Creating ilert alert source error: empty response")
 		return diag.Errorf("alert source response is empty")
 	}
 	d.SetId(strconv.FormatInt(result.AlertSource.ID, 10))
@@ -943,7 +943,7 @@ func resourceAlertSourceRead(ctx context.Context, d *schema.ResourceData, m inte
 			}
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for alert source with id '%s' to be read", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for alert source with id '%s' to be read, error: %s", d.Id(), err.Error()))
 			}
 			return resource.NonRetryableError(fmt.Errorf("could not read an alert source with ID %s, error: %s", d.Id(), err.Error()))
 		}
@@ -952,11 +952,12 @@ func resourceAlertSourceRead(ctx context.Context, d *schema.ResourceData, m inte
 	})
 
 	if err != nil {
+		log.Printf("[ERROR] Reading ilert alert source error: %s", err.Error())
 		return diag.FromErr(err)
 	}
 
 	if result == nil || result.AlertSource == nil {
-		log.Printf("[ERROR] Reading ilert alert source error: empty response ")
+		log.Printf("[ERROR] Reading ilert alert source error: empty response")
 		return diag.Errorf("alert source response is empty")
 	}
 
@@ -1148,7 +1149,7 @@ func resourceAlertSourceUpdate(ctx context.Context, d *schema.ResourceData, m in
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for alert source with id '%s' to be updated", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for alert source with id '%s' to be updated, error: %s", d.Id(), err.Error()))
 			}
 			return resource.NonRetryableError(fmt.Errorf("could not update an alert source with ID %s, error: %s", d.Id(), err.Error()))
 		}
@@ -1175,7 +1176,7 @@ func resourceAlertSourceDelete(ctx context.Context, d *schema.ResourceData, m in
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for alert source with id '%s' to be deleted", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for alert source with id '%s' to be deleted, error: %s", d.Id(), err.Error()))
 			}
 			return resource.NonRetryableError(fmt.Errorf("could not delete an alert source with ID %s, error: %s", d.Id(), err.Error()))
 		}
@@ -1220,6 +1221,7 @@ func resourceAlertSourceExists(d *schema.ResourceData, m interface{}) (bool, err
 	})
 
 	if err != nil {
+		log.Printf("[ERROR] Reading ilert alert source error: %s", err.Error())
 		return false, err
 	}
 	return result, nil