+++ ilert/resource_team.go
@@ -129,7 +129,7 @@ func resourceTeamCreate(ctx context.Context, d *schema.ResourceData, m interface
 		return diag.FromErr(err)
 	}
 	if result == nil || result.Team == nil {
-		log.Printf("[ERROR] Creating ilert team error: empty response ")
+		log.Printf("[ERROR] Creating ilert team error: empty response")
 		return diag.Errorf("team response is empty")
 	}
 
@@ -161,18 +161,19 @@ func resourceTeamRead(ctx context.Context, d *schema.ResourceData, m interface{}
 				time.Sleep(2 * time.Second)
 				return resource.RetryableError(fmt.Errorf("waiting for team to be read, error: %s", err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not read an team with ID %s", d.Id()))
+			return resource.NonRetryableError(fmt.Errorf("could not read an team with ID %s, error: %s", d.Id(), err.Error()))
 		}
 		result = r
 		return nil
 	})
 
 	if err != nil {
+		log.Printf("[ERROR] Reading ilert team error: %s", err.Error())
 		return diag.FromErr(err)
 	}
 
 	if result == nil || result.Team == nil {
-		log.Printf("[ERROR] Reading ilert team error: empty response ")
+		log.Printf("[ERROR] Reading ilert team error: empty response")
 		return diag.Errorf("team response is empty")
 	}
 
@@ -211,9 +212,9 @@ func resourceTeamUpdate(ctx context.Context, d *schema.ResourceData, m interface
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for team with id '%s' to be updated", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for team with id '%s' to be updated, error: %s", d.Id(), err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not update an team with ID %s", d.Id()))
+			return resource.NonRetryableError(fmt.Errorf("could not update an team with ID %s, error: %s", d.Id(), err.Error()))
 		}
 		return nil
 	})
@@ -240,9 +241,9 @@ func resourceTeamDelete(ctx context.Context, d *schema.ResourceData, m interface
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for team with id '%s' to be deleted", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for team with id '%s' to be deleted, error: %s", d.Id(), err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not delete an team with ID %s", d.Id()))
+			return resource.NonRetryableError(fmt.Errorf("could not delete an team with ID %s, error: %s", d.Id(), err.Error()))
 		}
 		return nil
 	})
@@ -278,13 +279,14 @@ func resourceTeamExists(d *schema.ResourceData, m interface{}) (bool, error) {
 				time.Sleep(2 * time.Second)
 				return resource.RetryableError(fmt.Errorf("waiting for team to be read, error: %s", err.Error()))
 			}
-			return resource.NonRetryableError(err)
+			return resource.NonRetryableError(fmt.Errorf("could not read a team with ID %s, error: %s", d.Id(), err.Error()))
 		}
 		result = true
 		return nil
 	})
 
 	if err != nil {
+		log.Printf("[ERROR] Reading ilert team error: %s", err.Error())
 		return false, err
 	}
 	return result, nil