+++ ilert/resource_user_email_contact.go
@@ -96,7 +96,7 @@ func resourceUserEmailContactCreate(ctx context.Context, d *schema.ResourceData,
 				time.Sleep(2 * time.Second)
 				return resource.RetryableError(fmt.Errorf("waiting for user email contact to be created, error: %s", err.Error()))
 			}
-			return resource.NonRetryableError(err)
+			return resource.NonRetryableError(fmt.Errorf("could not create a user email contact with ID %s, error: %s", d.Id(), err.Error()))
 		}
 		result = r
 		return nil
@@ -106,7 +106,7 @@ func resourceUserEmailContactCreate(ctx context.Context, d *schema.ResourceData,
 		return diag.FromErr(err)
 	}
 	if result == nil || result.UserEmailContact == nil {
-		log.Printf("[ERROR] Creating ilert user email contact error: empty response ")
+		log.Printf("[ERROR] Creating ilert user email contact error: empty response")
 		return diag.Errorf("user email contact response is empty")
 	}
 
@@ -147,20 +147,21 @@ func resourceUserEmailContactRead(ctx context.Context, d *schema.ResourceData, m
 			}
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for user email contact with id '%s' to be read", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for user email contact with id '%s' to be read, error: %s", d.Id(), err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not read an user email contact with id %s", d.Id()))
+			return resource.NonRetryableError(fmt.Errorf("could not read a user email contact with id %s, error: %s", d.Id(), err.Error()))
 		}
 		result = r
 		return nil
 	})
 
 	if err != nil {
+		log.Printf("[ERROR] Reading ilert user email contact error: %s", err.Error())
 		return diag.FromErr(err)
 	}
 
 	if result == nil || result.UserEmailContact == nil {
-		log.Printf("[ERROR] Reading ilert user email contact error: empty response ")
+		log.Printf("[ERROR] Reading ilert user email contact error: empty response")
 		return diag.Errorf("user email contact response is empty")
 	}
 
@@ -196,9 +197,9 @@ func resourceUserEmailContactUpdate(ctx context.Context, d *schema.ResourceData,
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for user email contact with id '%s' to be updated", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for user email contact with id '%s' to be updated, error: %s", d.Id(), err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not update an user email contact with id %s", d.Id()))
+			return resource.NonRetryableError(fmt.Errorf("could not update a user email contact with id %s, error: %s", d.Id(), err.Error()))
 		}
 		return nil
 	})
@@ -231,9 +232,9 @@ func resourceUserEmailContactDelete(ctx context.Context, d *schema.ResourceData,
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for user email contact with id '%s' to be deleted", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for user email contact with id '%s' to be deleted, error: %s", d.Id(), err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not delete an user email contact with id %s", d.Id()))
+			return resource.NonRetryableError(fmt.Errorf("could not delete a user email contact with id %s, error: %s", d.Id(), err.Error()))
 		}
 		return nil
 	})
@@ -275,13 +276,14 @@ func resourceUserEmailContactExists(d *schema.ResourceData, m interface{}) (bool
 				time.Sleep(2 * time.Second)
 				return resource.RetryableError(fmt.Errorf("waiting for user email contact to be read, error: %s", err.Error()))
 			}
-			return resource.NonRetryableError(err)
+			return resource.NonRetryableError(fmt.Errorf("could not read a user email contact with ID %s, error: %s", d.Id(), err.Error()))
 		}
 		result = true
 		return nil
 	})
 
 	if err != nil {
+		log.Printf("[ERROR] Reading ilert user email contact error: %s", err.Error())
 		return false, err
 	}
 	return result, nil