+++ ilert/resource_user_subscription_preference.go
@@ -124,7 +124,7 @@ func resourceUserSubscriptionPreferenceCreate(ctx context.Context, d *schema.Res
 				time.Sleep(2 * time.Second)
 				return resource.RetryableError(fmt.Errorf("waiting for user subscription preference to be created, error: %s", err.Error()))
 			}
-			return resource.NonRetryableError(err)
+			return resource.NonRetryableError(fmt.Errorf("could not create a user subscription preference with ID %s, error: %s", d.Id(), err.Error()))
 		}
 		result = r
 		return nil
@@ -134,7 +134,7 @@ func resourceUserSubscriptionPreferenceCreate(ctx context.Context, d *schema.Res
 		return diag.FromErr(err)
 	}
 	if result == nil || result.UserSubscriptionPreference == nil {
-		log.Printf("[ERROR] Creating ilert user subscription preference error: empty response ")
+		log.Printf("[ERROR] Creating ilert user subscription preference error: empty response")
 		return diag.Errorf("user subscription preference response is empty")
 	}
 
@@ -175,20 +175,21 @@ func resourceUserSubscriptionPreferenceRead(ctx context.Context, d *schema.Resou
 			}
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for user subscription preference with id '%s' to be read", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for user subscription preference with id '%s' to be read, error: %s", d.Id(), err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not read an user subscription preference with id %s", d.Id()))
+			return resource.NonRetryableError(fmt.Errorf("could not read a user subscription preference with id %s, error: %s", d.Id(), err.Error()))
 		}
 		result = r
 		return nil
 	})
 
 	if err != nil {
+		log.Printf("[ERROR] Reading ilert user subscription preference error: %s", err.Error())
 		return diag.FromErr(err)
 	}
 
 	if result == nil || result.UserSubscriptionPreference == nil {
-		log.Printf("[ERROR] Reading ilert user subscription preference error: empty response ")
+		log.Printf("[ERROR] Reading ilert user subscription preference error: empty response")
 		return diag.Errorf("user subscription preference response is empty")
 	}
 
@@ -231,9 +232,9 @@ func resourceUserSubscriptionPreferenceUpdate(ctx context.Context, d *schema.Res
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for user subscription preference with id '%s' to be updated", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for user subscription preference with id '%s' to be updated, error: %s", d.Id(), err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not update an user subscription preference with id %s", d.Id()))
+			return resource.NonRetryableError(fmt.Errorf("could not update a user subscription preference with id %s, error: %s", d.Id(), err.Error()))
 		}
 		return nil
 	})
@@ -266,9 +267,9 @@ func resourceUserSubscriptionPreferenceDelete(ctx context.Context, d *schema.Res
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for user subscription preference with id '%s' to be deleted", d.Id()))
+				return resource.RetryableError(fmt.Errorf("waiting for user subscription preference with id '%s' to be deleted, error: %s", d.Id(), err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not delete an user subscription preference with id %s", d.Id()))
+			return resource.NonRetryableError(fmt.Errorf("could not delete a user subscription preference with id %s, error: %s", d.Id(), err.Error()))
 		}
 		return nil
 	})
@@ -310,13 +311,14 @@ func resourceUserSubscriptionPreferenceExists(d *schema.ResourceData, m interfac
 				time.Sleep(2 * time.Second)
 				return resource.RetryableError(fmt.Errorf("waiting for user subscription preference to be read, error: %s", err.Error()))
 			}
-			return resource.NonRetryableError(err)
+			return resource.NonRetryableError(fmt.Errorf("could not read a user subscription preference with ID %s, error: %s", d.Id(), err.Error()))
 		}
 		result = true
 		return nil
 	})
 
 	if err != nil {
+		log.Printf("[ERROR] Reading ilert user subscription preference error: %s", err.Error())
 		return false, err
 	}
 	return result, nil